+++ CHANGES.rst
@@ -13,6 +13,7 @@ Improvements:
  - Support specifying kick and ban message (#2164)
  - Add image transparency and fix issues with gifs in the media viewer (#2731)
  - Migration to the Preference v7 support
+ - Make User Agreement part of the registration flow (#2442)
 
 Other changes:
  -